refactor(api-reference): update import paths and remove unused module declaration
