--- packages/api-reference/global.d.ts
@@ -1,4 +1,0 @@
-declare module '*.txt' {
-  const content: string
-  export default content
-}
